when delete user card, delete all data - INSA requests (Kristifor)

+++ src/main/web/src/app/pages/device/device.component.ts
@@ -576,13 +576,15 @@ export class DeviceComponent extends AbstractComponent implements OnInit {
   }
 
   delete() {
-    this.userCardService
+    if (confirm(this.translate.instant("Are you sure?"))) {
+      this.userCardService
       .deleteUserCard(this.device)
       .pipe(takeUntil(this.destroy$))
       .subscribe((val) => {
         this.displayDialog = false;
         this.loadPageable();
       });
+    }
   }
 
   loadStaticData() {
